+++ libs/lpplib/src/lpp.cpp
@@ -19,6 +19,7 @@ LPP_Client::LPP_Client(bool segmentation) {
     mSUPL                     = std::make_unique<SUPL_Client>();
     mSuplIdentityFix          = true;
     mLocationUpdateUnlocked   = false;
+    mLocationUpdateRate       = 1000;
 
     main_request_callback  = nullptr;
     main_request_userdata  = nullptr;
@@ -374,7 +375,8 @@ bool LPP_Client::process() {
             update_interval =
                 duration_cast<seconds>(std::chrono::milliseconds(provide_li.interval));
         } else {
-            update_interval = std::chrono::seconds(1);
+            update_interval =
+                duration_cast<seconds>(std::chrono::milliseconds(mLocationUpdateRate));
         }
 
         if (duration > update_interval) {
@@ -650,4 +652,8 @@ void LPP_Client::force_location_information() {
 
 void LPP_Client::unlock_update_rate() {
     mLocationUpdateUnlocked = true;
+}
+
+void LPP_Client::set_update_rate(int rate) {
+    mLocationUpdateRate = rate;
 }